Refactor helper functions into their own separate helper file
