+++ graphql_compiler/compiler/emit_cypher.py
@@ -2,52 +2,11 @@
 """Convert lowered IR basic blocks to Cypher query strings."""
 from itertools import chain
 
+from graphql_compiler.compiler import cypher_helpers
 from graphql_compiler.compiler.cypher_query import CypherStep
-from graphql_compiler.compiler.helpers import FoldScopeLocation, Location
+from graphql_compiler.compiler.helpers import FoldScopeLocation
 
 from .blocks import Fold, QueryRoot, Recurse, Traverse
-
-
-def get_fold_scope_location_full_path_name(fold_scope_location):
-    """Return a unique name with the full traversal path to this FoldScopeLocation."""
-    # HACK(Leon): Get a unique name for each vertex in a fold traversal in Cypher.
-    # For FoldScopeLocation objects, get_location_name() only uses the first edge in the traversal,
-    # which doesn't work for Cypher because we need to explicitly label every intermediate vertex
-    # along that path.
-    # For other query languages like MATCH and Gremlin, the fold directive doesn't require
-    # us to name all the intermediate vertices on the path, which is why this is Cypher-specific.
-    if not (isinstance(fold_scope_location, FoldScopeLocation)):
-        raise TypeError(u'Expected cypher_step.as_block.location to be of type '
-                        u'FoldScopeLocation. Instead, got object {} of type {}.'
-                        .format(fold_scope_location, type(fold_scope_location)))
-    base_location = fold_scope_location.base_location
-    base_query_path = base_location.query_path  # the path traversed so far
-    fold_path = fold_scope_location.fold_path  # the path specified at or within the folded scope.
-    full_path = base_query_path + tuple(u'_'.join(edge_name) for edge_name in fold_path)
-    location = Location(full_path, field=base_location.field,
-                        visit_counter=base_location.visit_counter)
-    if base_location.field is not None:
-        raise ValueError(u'Expected base_location\'s field to be None since this method is used to '
-                         u'traverse vertices for a fold scope and at no point do we navigate to a '
-                         u'field. However, field was {}'.format(base_location.field))
-    step_location_name, _ = location.get_location_name()
-    return step_location_name
-
-
-def get_unique_vertex_name_from_location(location):
-    """Return a unique name for this location, whether or not it's in a fold scope."""
-    if isinstance(location, FoldScopeLocation):
-        return get_fold_scope_location_full_path_name(location)
-    elif isinstance(location, Location):
-        location_name, _ = location.get_location_name()
-        return location_name
-    raise TypeError(u'Expected location to be of type Location or FoldScopeLocation. Instead got '
-                    u'type {} for location {}'.format(type(location), location))
-
-
-def get_collected_vertex_list_name(full_path_name):
-    """Return the name of the list generated by folding vertices with name full_path_name."""
-    return u'collected_' + full_path_name
 
 
 def _emit_code_from_cypher_step(cypher_step):
@@ -72,7 +31,7 @@ def _emit_code_from_cypher_step(cypher_step):
                              u'section. {}'.format(cypher_step))
 
     step_location = cypher_step.as_block.location  # destination vertex for current step's traversal
-    step_location_name = get_unique_vertex_name_from_location(step_location)
+    step_location_name = cypher_helpers.get_unique_vertex_name_from_location(step_location)
 
     is_fold_step = isinstance(step_location, FoldScopeLocation)
 
@@ -93,7 +52,8 @@ def _emit_code_from_cypher_step(cypher_step):
             u'%(left_edge_mark)s-[:%(edge_type)s%(quantifier)s]-%(right_edge_mark)s' +
             step_vertex_pattern
         )
-        linked_location_name = get_unique_vertex_name_from_location(cypher_step.linked_location)
+        linked_location_name = cypher_helpers.get_unique_vertex_name_from_location(
+            cypher_step.linked_location)
         template_data['linked_location'] = linked_location_name
 
     if has_where_block:
@@ -140,13 +100,9 @@ def _emit_with_clause_components(cypher_steps):
     location_names = set()
     for cypher_step in cypher_steps:
         location = cypher_step.as_block.location
-        location_name = get_unique_vertex_name_from_location(location)
+        location_name = cypher_helpers.get_unique_vertex_name_from_location(location)
         if isinstance(location, FoldScopeLocation):
-            location_name = get_collected_vertex_list_name(location_name)
-        else:
-            raise TypeError(u'Expected cypher_step\'s as_block\'s location to be either of type '
-                            u'Location or FoldScopeLocation. Instead, got type {} for location {}'
-                            .format(type(location), location))
+            location_name = cypher_helpers.get_collected_vertex_list_name(location_name)
         location_names.add(location_name)
 
     # Sort the locations, to ensure a deterministic order.
@@ -175,8 +131,9 @@ def _emit_with_clause_components_for_current_fold_scope(current_fold_scope_cyphe
                             .format(cypher_step, type(cypher_step),
                                     current_fold_scope_cypher_steps))
         fold_scope_location = cypher_step.as_block.location
-        full_vertex_name = get_fold_scope_location_full_path_name(fold_scope_location)
-        collected_name = get_collected_vertex_list_name(full_vertex_name)
+        full_vertex_name = cypher_helpers.get_fold_scope_location_full_path_name(
+            fold_scope_location)
+        collected_name = cypher_helpers.get_collected_vertex_list_name(full_vertex_name)
         vertex_names[u'collect(' + full_vertex_name + ')'] = collected_name
 
     # Sort the locations, to ensure a deterministic order.
